refactor: cleaned files over 300 files
