+++ src/app/admin/website-content/_components/WebsiteContentManagementPage.tsx
@@ -1,113 +1,11 @@
 "use client";
 
-import { Dialog as DialogPrimitive } from "@base-ui/react/dialog";
-import {
-  Building2,
-  Compass,
-  Images,
-  Landmark,
-  Sparkles,
-  Target,
-  Users,
-  XIcon,
-} from "lucide-react";
-import { useMemo, useState } from "react";
-import { toast } from "sonner";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "@/components/ui/alert-dialog";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Skeleton } from "@/components/ui/skeleton";
-import {
-  PendingUploadsProvider,
-  usePendingUploadsContext,
-} from "../_context/PendingUploadsContext";
-import { useWebsiteContentEditor } from "../_hooks/useWebsiteContentEditor";
-import { BoardOfTrusteesSection } from "./sections/BoardOfTrusteesSection";
-import { CompanyThrustsSection } from "./sections/CompanyThrustsSection";
-import { GoalsSection } from "./sections/GoalsSection";
-import { HeroSectionCarouselSection } from "./sections/HeroSectionCarouselSection";
-import { LandingBenefitsSection } from "./sections/LandingBenefitsSection";
-import { SecretariatSection } from "./sections/SecretariatSection";
-import { VisionMissionSection } from "./sections/VisionMissionSection";
-
-type SectionKey =
-  | "vision_mission"
-  | "goals"
-  | "company_thrusts"
-  | "board_of_trustees"
-  | "secretariat"
-  | "landing_page_benefits"
-  | "hero_section";
-
-const sectionCards = [
-  {
-    key: "vision_mission" as const,
-    title: "Vision and Mission",
-    description: "Paragraph content for organization vision and mission.",
-    accentClass: "bg-sky-500",
-    iconClass: "bg-sky-50 text-sky-600",
-    icon: Compass,
-  },
-  {
-    key: "goals" as const,
-    title: "Goals",
-    description: "Title, paragraph, and icon setup for goals section cards.",
-    accentClass: "bg-emerald-500",
-    iconClass: "bg-emerald-50 text-emerald-600",
-    icon: Target,
-  },
-  {
-    key: "company_thrusts" as const,
-    title: "Company Thrusts",
-    description: "Title, paragraph, and icon values for thrust items.",
-    accentClass: "bg-amber-500",
-    iconClass: "bg-amber-50 text-amber-600",
-    icon: Building2,
-  },
-  {
-    key: "board_of_trustees" as const,
-    title: "Board of Trustees",
-    description: "Manage title, subtitle, image, and optional card placement.",
-    accentClass: "bg-rose-500",
-    iconClass: "bg-rose-50 text-rose-600",
-    icon: Landmark,
-  },
-  {
-    key: "secretariat" as const,
-    title: "Secretariat",
-    description: "Manage title, subtitle, and image for staff cards.",
-    accentClass: "bg-cyan-500",
-    iconClass: "bg-cyan-50 text-cyan-600",
-    icon: Users,
-  },
-  {
-    key: "landing_page_benefits" as const,
-    title: "Landing Page Benefits",
-    description: "Set title, paragraph copy, and icon for home benefits.",
-    accentClass: "bg-indigo-500",
-    iconClass: "bg-indigo-50 text-indigo-600",
-    icon: Sparkles,
-  },
-  {
-    key: "hero_section" as const,
-    title: "Hero Section Carousel",
-    description:
-      "Manage 5 background carousel images each for About, Events, Members, Networks, and Contact pages.",
-    accentClass: "bg-violet-500",
-    iconClass: "bg-violet-50 text-violet-600",
-    icon: Images,
-    fields: "25 image slots",
-  },
-];
+import { PendingUploadsProvider } from "../_context/PendingUploadsContext";
+import { useWebsiteContentManagementPage } from "../_hooks/useWebsiteContentManagementPage";
+import { DeleteSelectedCardsDialog } from "./DeleteSelectedCardsDialog";
+import { WebsiteContentEditorDialog } from "./WebsiteContentEditorDialog";
+import { WebsiteContentSectionCardsGrid } from "./WebsiteContentSectionCardsGrid";
+import { websiteContentSections } from "./websiteContentSections";
 
 export function WebsiteContentManagementPage() {
   return (
@@ -118,290 +16,27 @@ export function WebsiteContentManagementPage() {
 }
 
 function WebsiteContentManagementPageContent() {
-  const [activeSection, setActiveSection] = useState<SectionKey | null>(null);
-  const [isUploadingImages, setIsUploadingImages] = useState(false);
-  const { uploadAllPendingImages } = usePendingUploadsContext();
-  const [isDeleteConfirmOpen, setIsDeleteConfirmOpen] = useState(false);
   const {
-    form,
-    cards,
-    placeholders,
-    setField,
-    setCardField,
-    replaceCards,
-    addCard,
-    isDeleteMode,
-    selectedCardEntryKeys,
-    enterDeleteMode,
-    cancelDeleteMode,
-    selectAllCards,
-    unselectAllCards,
-    toggleCardSelected,
-    deleteSelectedCards,
-    save,
-    isSavingSection,
-    saveSucceededCount,
+    activeSectionContent,
+    editingFooter,
+    handleConfirmDeleteSelectedCards,
+    handleEditorOpenChange,
+    handleSave,
+    hasSelectedCards,
+    isDeleteConfirmOpen,
+    isEditorOpen,
     isLoadingSection,
-    updatedAtBySection,
-    cardCountBySection,
-    hasLoadedSectionSummaries,
-  } = useWebsiteContentEditor(activeSection);
-
-  const [editingFooter, setEditingFooter] = useState<
-    { label: string; onClick: () => void } | undefined
-  >(undefined);
-
-  const hasSelectedCards = selectedCardEntryKeys.size > 0;
-  const selectedCount = selectedCardEntryKeys.size;
-  const isSectionActionDisabled = isSavingSection || isLoadingSection;
-  const isSaveDisabled =
-    isSavingSection || isLoadingSection || isDeleteMode || isUploadingImages;
-  const showCloseButton = !editingFooter;
-
-  const handleDeleteCardsClick = (entryKey?: string) => {
-    if (entryKey) {
-      setIsDeleteConfirmOpen(true);
-      if (!selectedCardEntryKeys.has(entryKey) || selectedCount !== 1) {
-        // Replace any existing selection so a card-level trash click deletes only that card.
-        unselectAllCards();
-        toggleCardSelected(entryKey, true);
-      }
-      return;
-    }
-
-    if (!isDeleteMode) {
-      enterDeleteMode();
-      return;
-    }
-
-    if (!hasSelectedCards) {
-      return;
-    }
-
-    setIsDeleteConfirmOpen(true);
-  };
-
-  const selectedCard = useMemo(
-    () => sectionCards.find((section) => section.key === activeSection) ?? null,
-    [activeSection],
-  );
-
-  const handleSave = async () => {
-    const snapshotForm = form;
-    const snapshotCards = cards;
-    let finalCards = snapshotCards;
-
-    setIsUploadingImages(true);
-    try {
-      const uploadedImages = await uploadAllPendingImages();
-      finalCards = snapshotCards.map((card) => {
-        const publicUrl = uploadedImages[card.entryKey];
-
-        if (!publicUrl) {
-          return card;
-        }
-
-        return {
-          ...card,
-          imageUrl: publicUrl,
-        };
-      });
-
-      const hasPreviewUrls = finalCards.some((card) => {
-        const imageUrl = card.imageUrl.trim().toLowerCase();
-        return imageUrl.startsWith("blob:") || imageUrl.startsWith("data:");
-      });
-
-      if (hasPreviewUrls) {
-        console.warn(
-          `Images still have preview URLs after upload: ${finalCards
-            .filter((c) => {
-              const imageUrl = c.imageUrl.trim().toLowerCase();
-              return (
-                imageUrl.startsWith("blob:") || imageUrl.startsWith("data:")
-              );
-            })
-            .map((c) => c.entryKey)
-            .join(", ")}`,
-        );
-        toast.error(
-          "Image upload is taking too long. Please check your connection and try again.",
-        );
-        return;
-      }
-    } catch (error) {
-      console.error("Error uploading images:", error);
-      toast.error("Failed to upload images. Please try again.");
-      return;
-    } finally {
-      setIsUploadingImages(false);
-    }
-
-    await save(finalCards, snapshotForm);
-  };
-
-  const updatedAtDisplay = (section: SectionKey) => {
-    if (!hasLoadedSectionSummaries) {
-      return null;
-    }
-
-    const updatedAt = updatedAtBySection[section];
-    if (!updatedAt) {
-      return "Not updated yet";
-    }
-
-    return new Date(updatedAt).toLocaleString();
-  };
-
-  const savedCardsDisplay = (section: SectionKey) => {
-    if (!hasLoadedSectionSummaries) {
-      return null;
-    }
-
-    return String(cardCountBySection[section] ?? 0);
-  };
-
-  const renderActiveForm = () => {
-    if (!activeSection) {
-      return null;
-    }
-
-    switch (activeSection) {
-      case "vision_mission":
-        return (
-          <VisionMissionSection
-            missionParagraph={form.missionParagraph}
-            onMissionParagraphChange={(value) =>
-              setField("missionParagraph", value)
-            }
-            onVisionParagraphChange={(value) =>
-              setField("visionParagraph", value)
-            }
-            placeholders={placeholders}
-            visionParagraph={form.visionParagraph}
-          />
-        );
-
-      case "goals":
-        return (
-          <GoalsSection
-            cards={cards}
-            hasSelectedCards={hasSelectedCards}
-            isDeleteMode={isDeleteMode}
-            isSectionActionDisabled={isSectionActionDisabled}
-            onAddCard={() => addCard()}
-            onCancelDeleteMode={cancelDeleteMode}
-            onCardFieldChange={setCardField}
-            onDeleteCardsClick={handleDeleteCardsClick}
-            onSelectAllCards={selectAllCards}
-            onToggleCardSelected={toggleCardSelected}
-            onUnselectAllCards={unselectAllCards}
-            placeholders={placeholders}
-            selectedCardEntryKeys={selectedCardEntryKeys}
-            selectedCount={selectedCount}
-          />
-        );
-
-      case "company_thrusts":
-        return (
-          <CompanyThrustsSection
-            cards={cards}
-            hasSelectedCards={hasSelectedCards}
-            isDeleteMode={isDeleteMode}
-            isSectionActionDisabled={isSectionActionDisabled}
-            onAddCard={() => addCard()}
-            onCancelDeleteMode={cancelDeleteMode}
-            onCardFieldChange={setCardField}
-            onDeleteCardsClick={handleDeleteCardsClick}
-            onSelectAllCards={selectAllCards}
-            onToggleCardSelected={toggleCardSelected}
-            onUnselectAllCards={unselectAllCards}
-            placeholders={placeholders}
-            selectedCardEntryKeys={selectedCardEntryKeys}
-            selectedCount={selectedCount}
-          />
-        );
-
-      case "board_of_trustees":
-        return (
-          <BoardOfTrusteesSection
-            cards={cards}
-            hasSelectedCards={hasSelectedCards}
-            isDeleteMode={isDeleteMode}
-            isSectionActionDisabled={isSectionActionDisabled}
-            onAddCard={(group) => addCard(group)}
-            onCancelDeleteMode={cancelDeleteMode}
-            onCardFieldChange={setCardField}
-            onCardsReorder={replaceCards}
-            onDeleteCardsClick={handleDeleteCardsClick}
-            onRegisterEditingFooter={setEditingFooter}
-            onSelectAllCards={selectAllCards}
-            onToggleCardSelected={toggleCardSelected}
-            onUnselectAllCards={unselectAllCards}
-            placeholders={placeholders}
-            saveSucceededCount={saveSucceededCount}
-            selectedCardEntryKeys={selectedCardEntryKeys}
-            selectedCount={selectedCount}
-          />
-        );
-
-      case "secretariat":
-        return (
-          <SecretariatSection
-            cards={cards}
-            hasSelectedCards={hasSelectedCards}
-            isDeleteMode={isDeleteMode}
-            isSavingSection={isSavingSection}
-            isSectionActionDisabled={isSectionActionDisabled}
-            onAddCard={() => addCard()}
-            onCancelDeleteMode={cancelDeleteMode}
-            onCardFieldChange={setCardField}
-            onCardsReorder={replaceCards}
-            onDeleteCardsClick={handleDeleteCardsClick}
-            onRegisterEditingFooter={setEditingFooter}
-            onSelectAllCards={selectAllCards}
-            onToggleCardSelected={toggleCardSelected}
-            onUnselectAllCards={unselectAllCards}
-            placeholders={placeholders}
-            saveSucceededCount={saveSucceededCount}
-            selectedCardEntryKeys={selectedCardEntryKeys}
-            selectedCount={selectedCount}
-          />
-        );
-
-      case "landing_page_benefits":
-        return (
-          <LandingBenefitsSection
-            cards={cards}
-            hasSelectedCards={hasSelectedCards}
-            isDeleteMode={isDeleteMode}
-            isSectionActionDisabled={isSectionActionDisabled}
-            onAddCard={() => addCard()}
-            onCancelDeleteMode={cancelDeleteMode}
-            onCardFieldChange={setCardField}
-            onDeleteCardsClick={handleDeleteCardsClick}
-            onSelectAllCards={selectAllCards}
-            onToggleCardSelected={toggleCardSelected}
-            onUnselectAllCards={unselectAllCards}
-            placeholders={placeholders}
-            selectedCardEntryKeys={selectedCardEntryKeys}
-            selectedCount={selectedCount}
-          />
-        );
-
-      case "hero_section":
-        return (
-          <HeroSectionCarouselSection
-            cards={cards}
-            onCardFieldChange={setCardField}
-            onCardsReorder={replaceCards}
-          />
-        );
-
-      default:
-        return null;
-    }
-  };
+    isSaveDisabled,
+    isSavingSection,
+    isSectionActionDisabled,
+    isUploadingImages,
+    openSection,
+    savedCardsDisplay,
+    selectedCard,
+    selectedCount,
+    setIsDeleteConfirmOpen,
+    updatedAtDisplay,
+  } = useWebsiteContentManagementPage();
 
   return (
     <section className="space-y-6 px-2">
@@ -414,182 +49,35 @@ function WebsiteContentManagementPageContent() {
         </p>
       </header>
 
-      <div className="grid gap-6 md:grid-cols-2 xl:grid-cols-3">
-        {sectionCards.map((section) => {
-          const Icon = section.icon;
+      <WebsiteContentSectionCardsGrid
+        getSavedCardsDisplay={savedCardsDisplay}
+        getUpdatedAtDisplay={updatedAtDisplay}
+        onOpenSection={openSection}
+        sections={websiteContentSections}
+      />
 
-          const savedCardsValue = savedCardsDisplay(section.key);
-          const updatedAtValue = updatedAtDisplay(section.key);
-
-          return (
-            <button
-              className="group text-left"
-              key={section.key}
-              onClick={() => setActiveSection(section.key)}
-              type="button"
-            >
-              <Card className="relative h-full overflow-hidden border border-border/80 bg-card/95 shadow-sm transition-all duration-200 group-hover:-translate-y-0.5 group-hover:shadow-md">
-                <div
-                  className={`absolute inset-x-0 top-0 h-1 ${section.accentClass}`}
-                />
-                <CardHeader>
-                  <div className="flex items-start gap-3">
-                    <div className={`rounded-xl p-2.5 ${section.iconClass}`}>
-                      <Icon className="h-5 w-5" />
-                    </div>
-                    <div className="space-y-1">
-                      <CardTitle className="text-xl">{section.title}</CardTitle>
-                      <p className="text-muted-foreground text-sm">
-                        {section.description}
-                      </p>
-                    </div>
-                  </div>
-                </CardHeader>
-                <CardContent className="space-y-2 text-muted-foreground text-sm">
-                  <div className="flex items-center justify-between gap-3">
-                    <span>Saved cards:</span>
-                    {savedCardsValue ? (
-                      <span className="tabular-nums">{savedCardsValue}</span>
-                    ) : (
-                      <Skeleton className="h-4 w-10" />
-                    )}
-                  </div>
-                  <div className="flex items-center justify-between gap-3">
-                    <span>Updated:</span>
-                    {updatedAtValue ? (
-                      <span className="truncate">{updatedAtValue}</span>
-                    ) : (
-                      <Skeleton className="h-4 w-28" />
-                    )}
-                  </div>
-                </CardContent>
-              </Card>
-            </button>
-          );
-        })}
-      </div>
-
-      <DialogPrimitive.Root
-        onOpenChange={(open) => {
-          if (!open) {
-            cancelDeleteMode();
-            setActiveSection(null);
-            setEditingFooter(undefined);
-          }
-        }}
-        open={!!activeSection}
+      <WebsiteContentEditorDialog
+        editingFooter={editingFooter}
+        isLoadingSection={isLoadingSection}
+        isSaveDisabled={isSaveDisabled}
+        isSavingSection={isSavingSection}
+        isSectionActionDisabled={isSectionActionDisabled}
+        isUploadingImages={isUploadingImages}
+        onOpenChange={handleEditorOpenChange}
+        onSave={handleSave}
+        open={isEditorOpen}
+        selectedCard={selectedCard}
       >
-        <DialogPrimitive.Portal>
-          <DialogPrimitive.Backdrop className="fixed inset-0 z-50 bg-black/35 backdrop-blur-[2px]" />
-          <DialogPrimitive.Viewport className="fixed inset-0 z-50 overflow-hidden p-3 sm:p-6">
-            <div className="flex h-full items-center justify-center">
-              <DialogPrimitive.Popup className="relative flex max-h-[calc(100vh-6rem)] w-[min(97vw,1550px)] flex-col overflow-hidden rounded-2xl border border-border bg-background shadow-2xl">
-                {showCloseButton ? (
-                  <DialogPrimitive.Close
-                    className="absolute top-4 right-4 inline-flex h-9 w-9 items-center justify-center rounded-md text-muted-foreground transition-colors hover:bg-muted hover:text-foreground"
-                    render={<button type="button" />}
-                  >
-                    <XIcon className="h-5 w-5" />
-                    <span className="sr-only">Close</span>
-                  </DialogPrimitive.Close>
-                ) : null}
+        {activeSectionContent}
+      </WebsiteContentEditorDialog>
 
-                <div className="space-y-1.5 border-border border-b px-6 py-5 pr-12 sm:px-7">
-                  <DialogPrimitive.Title className="font-bold text-4xl text-foreground">
-                    {selectedCard?.title}
-                  </DialogPrimitive.Title>
-                  <DialogPrimitive.Description className="text-base text-muted-foreground">
-                    {selectedCard?.description}
-                  </DialogPrimitive.Description>
-                </div>
-
-                <div className="flex-1 overflow-y-auto px-6 py-5 sm:px-7">
-                  <div className="space-y-4">
-                    {isLoadingSection ? (
-                      <div className="space-y-4">
-                        <div className="space-y-2">
-                          <Skeleton className="h-5 w-40" />
-                          <Skeleton className="h-4 w-full" />
-                          <Skeleton className="h-4 w-5/6" />
-                        </div>
-                        <div className="grid grid-cols-1 gap-6 lg:grid-cols-2">
-                          <div className="space-y-3 rounded-lg border p-4">
-                            <div className="flex items-center justify-between">
-                              <Skeleton className="h-4 w-24" />
-                              <Skeleton className="h-8 w-8 rounded-md" />
-                            </div>
-                            <Skeleton className="h-10 w-full" />
-                            <Skeleton className="h-10 w-full" />
-                            <Skeleton className="h-40 w-full rounded-xl" />
-                          </div>
-                          <div className="flex items-center justify-center rounded-lg border p-4">
-                            <Skeleton className="h-[220px] w-[180px] rounded-md" />
-                          </div>
-                        </div>
-                      </div>
-                    ) : (
-                      renderActiveForm()
-                    )}
-                  </div>
-                </div>
-
-                <div className="border-border border-t bg-background px-6 py-4 sm:px-7">
-                  <div className="flex flex-wrap items-center justify-end gap-2">
-                    {editingFooter ? (
-                      <Button
-                        disabled={isSectionActionDisabled}
-                        onClick={() => {
-                          // invoke the section-provided back handler
-                          editingFooter.onClick();
-                        }}
-                      >
-                        {editingFooter.label}
-                      </Button>
-                    ) : (
-                      <Button disabled={isSaveDisabled} onClick={handleSave}>
-                        {isUploadingImages
-                          ? "Uploading images..."
-                          : isSavingSection
-                            ? "Saving..."
-                            : "Save Changes"}
-                      </Button>
-                    )}
-                  </div>
-                </div>
-              </DialogPrimitive.Popup>
-            </div>
-          </DialogPrimitive.Viewport>
-        </DialogPrimitive.Portal>
-      </DialogPrimitive.Root>
-
-      <AlertDialog
+      <DeleteSelectedCardsDialog
+        hasSelectedCards={hasSelectedCards}
+        onConfirmDelete={handleConfirmDeleteSelectedCards}
         onOpenChange={setIsDeleteConfirmOpen}
         open={isDeleteConfirmOpen}
-      >
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>Delete selected cards?</AlertDialogTitle>
-            <AlertDialogDescription>
-              This will remove {selectedCardEntryKeys.size} selected card
-              {selectedCardEntryKeys.size === 1 ? "" : "s"} from the editor.
-              Click Save Changes to persist this deletion.
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>Cancel</AlertDialogCancel>
-            <AlertDialogAction
-              className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-              disabled={!hasSelectedCards}
-              onClick={() => {
-                deleteSelectedCards();
-                setIsDeleteConfirmOpen(false);
-              }}
-            >
-              Confirm Delete
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
+        selectedCount={selectedCount}
+      />
     </section>
   );
 }
