+++ src/server/website-content/queries/getPublicWebsiteContentSection.ts
@@ -8,6 +8,13 @@ import { CACHE_TAGS } from "@/lib/cache/tags";
 import { createClient } from "@/lib/supabase/server";
 import { websiteContentSectionSchema } from "../schemas";
 import type { WebsiteContentSection } from "../types";
+import {
+  getHeroSectionImages,
+  type HeroSectionPage,
+  normalizePersonalImageUrl,
+} from "./publicWebsiteContentSection.utils";
+
+export type { HeroSectionPage } from "./publicWebsiteContentSection.utils";
 
 export type PublicWebsiteContentCard = {
   entryKey: string;
@@ -27,42 +34,6 @@ export type PublicWebsiteContentSectionData = {
   missionParagraph: string;
   cards: PublicWebsiteContentCard[];
 };
-
-export type HeroSectionPage =
-  | "about"
-  | "events"
-  | "members"
-  | "networks"
-  | "contact";
-
-export function getHeroSectionImages(
-  sectionData: PublicWebsiteContentSectionData,
-  page: HeroSectionPage,
-  limit = 5,
-): string[] {
-  if (sectionData.section !== "hero_section") {
-    return [];
-  }
-
-  return sectionData.cards
-    .filter((card) => card.group === page && card.imageUrl.trim().length > 0)
-    .sort((a, b) => {
-      const aPlacement = a.cardPlacement
-        ? Number(a.cardPlacement)
-        : Number.MAX_SAFE_INTEGER;
-      const bPlacement = b.cardPlacement
-        ? Number(b.cardPlacement)
-        : Number.MAX_SAFE_INTEGER;
-
-      if (aPlacement === bPlacement) {
-        return a.entryKey.localeCompare(b.entryKey);
-      }
-
-      return aPlacement - bPlacement;
-    })
-    .slice(0, limit)
-    .map((card) => card.imageUrl);
-}
 
 export async function getPublicHeroSectionImages(
   page: HeroSectionPage,
@@ -98,32 +69,6 @@ type WebsiteContentRow = {
   cardPlacement: number | null;
   group: string | null;
 };
-
-function normalizePersonalImageUrl(imageUrl: string | null) {
-  if (!imageUrl) {
-    return "";
-  }
-
-  if (/^https?:\/\//i.test(imageUrl)) {
-    return imageUrl.replace("/headerimage/", "/personalimage/");
-  }
-
-  const cleaned = imageUrl.replace(/^\/+/, "");
-  if (cleaned.startsWith("storage/v1/object/")) {
-    const baseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-    if (!baseUrl) {
-      return "";
-    }
-    return `${baseUrl}/${cleaned.replace("/headerimage/", "/personalimage/")}`;
-  }
-
-  const baseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-  if (!baseUrl) {
-    return "";
-  }
-
-  return `${baseUrl}/storage/v1/object/public/personalimage/${cleaned}`;
-}
 
 async function getPublicWebsiteContentSectionCached(
   requestCookies: RequestCookie[],
